recall page fixed for update and delete investigations and prescriptions

+++ src/main/java/com/hims/mapper/InvestigationData.java
@@ -7,6 +7,6 @@ public record InvestigationData(
         String investigationName,
         LocalDate investigationDate,
         Long investigationId,
-        Long flag
+        Integer flag
 ) {
 }
